+++ keeperfx_vs2010.vcxproj
@@ -191,6 +191,7 @@
     <ClCompile Include="src\net_game.c" />
     <ClCompile Include="src\net_sync.c" />
     <ClCompile Include="src\packets.c" />
+    <ClCompile Include="src\packets_input.c" />
     <ClCompile Include="src\packets_misc.c" />
     <ClCompile Include="src\player_compchecks.c" />
     <ClCompile Include="src\player_compevents.c" />
